Add setup for user-authored password protection
Middleware and basic HTTP authentication provides the simplest implementation. We do however want the client to be able to turn password protection on/off,  so that adds some complexity. We could simply fetch the user settings from Sanity in a middleware, but that would introduce latency. We could cache the settings in Redis, but that could affect previewing content quickly. Therefore the chosen implementation is to use the Vercel API to update a `PASSWORD_PROTECTION` environment variable.

This commit introduces an API endpoint which will update the environment variable to contain the password protection configuration that the user authors.

A webhook is set up in Sanity to his this endpoint whenever the `privacy` document gets updated.